feat(multicrafter): 添加多配方支持和配方显示功能
- 修改 ContentDisplay 类的 amount 参数类型从 int 改为 float
- 在 MCMBlocks 中添加液体依赖项并增加三个测试配方
- 实现 MCMRecipeDisplay 类用于显示配方输入输出
- 在 MultiCrafter 类中添加 recipes 序列和 addRecipe 方法
- 更新 MultiCrafterEditorDialog 以接收和显示多个配方
- 移除原来的占位符代码并实现完整的配方列表显示

+++ src/mxyz1733mods/ui/ContentDisplay.java
@@ -13,17 +13,17 @@ import mindustry.ui.Styles;
 import mindustry.world.meta.StatValues;
 
 public class ContentDisplay extends Table {
-    public ContentDisplay(@Nullable UnlockableContent content, int amount) {
+    public ContentDisplay(@Nullable UnlockableContent content, float amount) {
         rebuild(content, amount);
     }
 
-    private void rebuild(@Nullable UnlockableContent content, int amount) {
+    private void rebuild(@Nullable UnlockableContent content, float amount) {
         clear();
         top().left().margin(0.0F);
         add(stack(content, amount, true));
     }
 
-    private static Stack stack(@Nullable UnlockableContent content, int amount, boolean tooltip) {
+    private static Stack stack(@Nullable UnlockableContent content, float amount, boolean tooltip) {
         Stack stack = new Stack();
         stack.add(new Table(o -> {
             o.left();
